requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
discord.py==2.3.2
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
idna==3.4
multidict==6.0.4
numpy==1.26.1
outcome==1.3.0
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.7
wsproto==1.2.0
yarl==1.9.2
